Stub out all the rpcap-URL-related APIs if no remote capture.
The code needs a lot of cleanup, due o a number of bugs found in it.

For now, at least, let's just stub it out, with routines that always
fail and provide a "this isn't supported" error string, if ENABLE_REMOTE
isn't defined.

Move routines back into pcap-new.c; they are routines provided for 1)
the ability to open and enumerate remote devices and 2) compatibility
with WinPcap for access to local devices and capture files, but those
aren't APIs we want to recommend.
